refactored

+++ HomeTask_WindowsForms/Forms/CategoriesManagement.cs
@@ -24,27 +24,19 @@ namespace HomeTask_WindowsForms
 
         private void PrepareForm()
         {
-            // asking new lists
-            LocalAppData.Categories = _categoryRepository.GetAllCategories().ToList();
-            LocalAppData.Words = _wordRepository.GetAllWords().ToList();
-
             buttonUpdateCategory.Enabled = false;
             buttonDeleteCategory.Enabled = false;
 
             textBoxNewCategoryName.Text = "enter new category name name here";
             textBoxNewCategoryName.ForeColor = Color.Gray;
 
-            DrawTable();
-        }
-
-        private void DrawTable()
-        {
             LocalAppData.CountWordsInCategories();
-            
-            bindingSourceCategoryManagement.DataSource = LocalAppData.Categories;
 
+            bindingSourceCategoryManagement.ResetBindings(true);
+            bindingSourceCategoryManagement.DataSource = LocalAppData.Categories;
+            
             dataGridViewCategoriesManagement.ClearSelection(); // remove selection from first row
-            //throw new NotImplementedException();
+            
         }
 
         void dataGridViewCategoriesManagement_CellClick(object sender, DataGridViewCellEventArgs e)
@@ -77,7 +69,11 @@ namespace HomeTask_WindowsForms
             if (!textBoxNewCategoryName.Text.Equals("") && !textBoxNewCategoryName.Text.Equals(" ") &&
                 !textBoxNewCategoryName.Text.Equals("enter new category/category name name here"))
             {
-                _categoryRepository.AddCategory(new Category(textBoxNewCategoryName.Text) );
+                var newCategory = new Category(textBoxNewCategoryName.Text);
+                
+                var newCategoryId = _categoryRepository.AddCategory(newCategory);
+                
+                LocalAppData.Categories.Add(new Category(newCategoryId, newCategory.CategoryName, newCategory.IsUsed));
 
                 PrepareForm();
             }
@@ -85,20 +81,37 @@ namespace HomeTask_WindowsForms
 
         private void buttonDeleteCategory_Click(object sender, EventArgs e)
         {
-            var categoryToDelete = (Category)dataGridViewCategoriesManagement.CurrentRow.DataBoundItem;
+            if (dataGridViewCategoriesManagement.CurrentRow != null)
+            {
+                var categoryToDelete = (Category) dataGridViewCategoriesManagement.CurrentRow.DataBoundItem;
 
-            _wordRepository.UpdateWordsCategory(categoryToDelete);   // setting free words category "no category"
-            _categoryRepository.RemoveCategory(categoryToDelete);  
-            
+                // setting free words category "no category"
+                _wordRepository.UpdateWordsCategory(categoryToDelete);
+                LocalAppData.UpdateCategoryInWordsWhileDeleting(categoryToDelete);
+
+                _categoryRepository.RemoveCategory(categoryToDelete);
+                LocalAppData.Categories.Remove(categoryToDelete);
+            }
+
             PrepareForm();
         }
 
         private void buttonUpdateCategory_Click(object sender, EventArgs e)
         {
-            var categoryToUpdate = (Category) dataGridViewCategoriesManagement.CurrentRow.DataBoundItem;
+            if (dataGridViewCategoriesManagement.CurrentRow != null)
+            {
+                var categoryToUpdate = (Category) dataGridViewCategoriesManagement.CurrentRow.DataBoundItem;
+                var indexOfCategoriesList = LocalAppData.Categories.IndexOf(categoryToUpdate);
 
-            _categoryRepository.UpdateCategory(categoryToUpdate);
-            
+                LocalAppData.UpdateCategoryInWords(categoryToUpdate);
+
+                // updating category
+                _categoryRepository.UpdateCategory(categoryToUpdate);
+
+                LocalAppData.Categories.RemoveAt(indexOfCategoriesList);
+                LocalAppData.Categories.Insert(indexOfCategoriesList, categoryToUpdate);
+            }
+
             PrepareForm();
         }
     }
